refactor encryption package
