+++ x/kms/kms.go
@@ -5,6 +5,7 @@ import (
 	b64 "encoding/base64"
 
 	awskms "github.com/aws/aws-sdk-go-v2/service/kms"
+	"github.com/cultureamp/ca-go/x/encryption"
 	"github.com/pkg/errors"
 )
 
@@ -13,21 +14,16 @@ type KMSClient interface {
 	Decrypt(ctx context.Context, params *awskms.DecryptInput, optFns ...func(*awskms.Options)) (*awskms.DecryptOutput, error)
 }
 
-type KMSEncrypt interface {
-	Encrypt(ctx context.Context, plainStr string) (encryptedStr *string, err error)
-	Decrypt(ctx context.Context, encryptedStr string) (decryptedStr *string, err error)
-}
-
 type kmsEncrypt struct {
 	client KMSClient
 	keyID  *string
 }
 
-func NewKMSWithClient(keyID string, client KMSClient) KMSEncrypt {
+func NewKMSWithClient(keyID string, client KMSClient) encryption.Encryptor {
 	return &kmsEncrypt{client, &keyID}
 }
 
-func NewKMS(keyID string) KMSEncrypt {
+func NewKMS(keyID string) encryption.Encryptor {
 	return &kmsEncrypt{nil, &keyID}
 }
